+++ Tests/ParamConverter/DataTransferObjectParamConverterTest.php
@@ -582,6 +582,16 @@ class DataTransferObjectParamConverterTest extends MockeryTestCase
             ->with(
                 \Mockery::type(DummyDataTransferObject::class),
                 null,
+                'dto_raw_input_validation'
+            )
+            ->andReturn(new ConstraintViolationList());
+
+        $this->validator
+            ->shouldReceive('validate')
+            ->once()
+            ->with(
+                \Mockery::type(DummyDataTransferObject::class),
+                null,
                 'Default'
             )
             ->andReturn(new ConstraintViolationList());